Updated Authentication. Added a round package and Deleted Database.java and Modified Db_connect.java
