+++ src/DatabaseManager/Db_connect.java
@@ -3,13 +3,16 @@ package DatabaseManager;
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.SQLException;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
 
 public class Db_connect {
     private static final String URL = "jdbc:postgresql://localhost:5432/your_database";
-    private static final String USER = "your_user";
-    private static final String PASSWORD = "password";
+    private static final String USER = "postgres";
+    private static final String PASSWORD = "your_password";
 
-    private Db_connect() {}
+    private Db_connect() {
+    }
 
     public static Connection getConnection() {
         try {
@@ -37,5 +40,33 @@ public class Db_connect {
         } else {
             System.out.println("❌ Failed to connect to the database.");
         }
+    }
+
+    public static boolean authenticate(String username, String password, String role) {
+        // SQL query to select the password for the given username and role
+        String sql = "SELECT password FROM users WHERE username = ? AND role = ?";
+
+        // Use try-with-resources to ensure resources are closed automatically
+        try (Connection conn = getConnection(); // Establish a connection
+             PreparedStatement pstmt = conn.prepareStatement(sql)) { // Prepare the SQL statement
+
+            // Set the username and role parameters in the SQL query
+            pstmt.setString(1, username);
+            pstmt.setString(2, role); // Set the role parameter
+
+            ResultSet rs = pstmt.executeQuery(); // Execute the query
+
+            // Check if a result is returned
+            if (rs.next()) {
+                // Retrieve the stored password from the result set
+                String storedPassword = rs.getString("password");
+                // Compare the stored password with the provided password
+                return storedPassword.equals(password); // Replace with password hashing if needed
+            }
+        } catch (Exception e) {
+            // Print the stack trace if an error occurs
+            e.printStackTrace();
+        }
+        return false; // Return false if authentication fails
     }
 }
