+++ src/Round/RoundedPasswordField.java
@@ -1,0 +1,30 @@
+package Round;
+
+import javax.swing.*;
+import java.awt.*;
+
+public class RoundedPasswordField extends JPasswordField {
+    private int radius;
+
+    public RoundedPasswordField(int columns, int radius) {
+        super(columns);
+        this.radius = radius;
+        setOpaque(false); // Make background transparent
+        setBorder(BorderFactory.createEmptyBorder(5, 10, 5, 10));
+    }
+
+    protected void paintComponent(Graphics g) {
+        Graphics2D g2 = (Graphics2D) g.create();
+        g2.setRenderingHint(RenderingHints.KEY_ANTIALIASING, RenderingHints.VALUE_ANTIALIAS_ON);
+
+        g2.setColor(Color.WHITE);
+        g2.fillRoundRect(0, 0, getWidth(), getHeight(), radius, radius);
+
+        g2.setColor(Color.GRAY);
+        g2.drawRoundRect(0, 0, getWidth() - 1, getHeight() - 1, radius, radius);
+
+        g2.dispose();
+        super.paintComponent(g);
+    }
+}
+
